Better-organize spec helpers
