+++ spec/support/preload.rb
@@ -1,6 +1,6 @@
 # List all helpers here that should be required in spec_helper.rb.
 %w[
-  upsteem_spec_helper_loader
+  spec_helper_loader
 ].each do |file|
-  require_relative(file)
+  require_relative("upsteem/deploy/#{file}")
 end
