+++ spec/unit/upsteem/deploy/deployer_spec.rb
@@ -1,5 +1,5 @@
 require "spec_helper"
-Upsteem::SpecHelperLoader.require_shared_examples_for("exception_raiser")
+Upsteem::Deploy::SpecHelperLoader.require_shared_examples_for("exception_raiser")
 
 describe Upsteem::Deploy::Deployer do
   let(:task_class) { class_double("Upsteem::Deploy::Tasks::Task") }
