+++ spec/unit/upsteem/deploy/environment_spec.rb
@@ -1,5 +1,5 @@
 require "spec_helper"
-Upsteem::SpecHelperLoader.require_shared_examples_for("exception_raiser")
+Upsteem::Deploy::SpecHelperLoader.require_shared_examples_for("exception_raiser")
 
 describe Upsteem::Deploy::Environment do
   let(:configuration) { instance_double("Upsteem::Deploy::Configuration") }
